+++ cpp/src/join/conditional_join.cu
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include <cudf/ast/nodes.hpp>
+#include <cudf/ast/expressions.hpp>
 #include <cudf/detail/utilities/cuda.cuh>
 #include <cudf/join.hpp>
 #include <cudf/table/table.hpp>
