+++ cpp/src/join/conditional_join.hpp
@@ -17,7 +17,7 @@
 
 #include "join_common_utils.hpp"
 
-#include <cudf/ast/nodes.hpp>
+#include <cudf/ast/expressions.hpp>
 #include <cudf/table/table_view.hpp>
 
 #include <rmm/cuda_stream_view.hpp>
