+++ cpp/src/transform/compute_column.cu
@@ -16,7 +16,7 @@
 
 #include <cudf/ast/detail/evaluator.cuh>
 #include <cudf/ast/detail/linearizer.hpp>
-#include <cudf/ast/nodes.hpp>
+#include <cudf/ast/expressions.hpp>
 #include <cudf/column/column_device_view.cuh>
 #include <cudf/column/column_factories.hpp>
 #include <cudf/detail/nvtx/ranges.hpp>
